Enable or Disable decrement animation
Currently I'm reusing a progress bar for a multiple steps process, the decrement animation feels a bit weird like "undoing". With this change by default you don't get this animation. On increment the animation shows cool, but if the next value is lower it meens a new process as started so it starts from that value without animation. There is a new bool property AnimatedProgressDecrease and you can set it to true if you want the effect in both ways.